move admin controller to user controller
